requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.0
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
execnet==2.0.2
idna==3.6
iniconfig==2.0.0
model-bakery==1.17.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
pytest-xdist==3.5.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
urllib3==2.1.0
